# apps/gateway/internal/modules/auth/dto/refresh.dto.go
package dto

import (
	"fmt"

	"github.com/go-playground/validator/v10"
)

// RefreshTokenRequest represents refresh token request
// @Description Refresh token request payload
type RefreshTokenRequest struct {
	// Refresh token
	// @Example eyJhbGciOiJIUzI1...
	RefreshToken string `json:"refresh_token" validate:"required" example:"eyJhbGciOiJIUzI1..."`
}

func (c *RefreshTokenRequest) Validate() error {
	validate := validator.New()
	err := validate.Struct(c)
	if err != nil {
		if _, ok := err.(*validator.InvalidValidationError); ok {
			return err
		}
		var errorMessage string
		for _, err := range err.(validator.ValidationErrors) {
			errorMessage += fmt.Sprintf("Field: %s, Error: %s, Value: %v\n",
				err.Field(),
				err.Tag(),
				err.Value())
		}
		return fmt.Errorf(errorMessage)
	}
	return nil
}